minor: changed powermock tests to normal tests for TranslationCheck
